+++ networkInfo.go
@@ -5,7 +5,7 @@ import (
 	"strings"
 )
 
-func (s *Ip) GetNumberIPAddresses() int {
+func (s *Ip) GetNumberIPAddresses() int64 {
 	return 2 << uint(31-s.networkSize)
 }
 
@@ -30,10 +30,10 @@ func (s *Ip) GetBroadcastAddress() string {
 	networkQuads := s.GetNetworkPortionQuards()
 	numberIPAddress := s.GetNumberIPAddresses()
 	networkRangeQuads := []string{}
-	networkRangeQuads = append(networkRangeQuads, fmt.Sprintf("%d", (networkQuads[0]&(s.subnet_mask>>24))+(((numberIPAddress-1)>>24)&0xFF)))
-	networkRangeQuads = append(networkRangeQuads, fmt.Sprintf("%d", (networkQuads[1]&(s.subnet_mask>>16))+(((numberIPAddress-1)>>16)&0xFF)))
-	networkRangeQuads = append(networkRangeQuads, fmt.Sprintf("%d", (networkQuads[2]&(s.subnet_mask>>8))+(((numberIPAddress-1)>>8)&0xFF)))
-	networkRangeQuads = append(networkRangeQuads, fmt.Sprintf("%d", (networkQuads[3]&(s.subnet_mask>>0))+(((numberIPAddress-1)>>0)&0xFF)))
+	networkRangeQuads = append(networkRangeQuads, fmt.Sprintf("%d", (networkQuads[0]&(s.subnet_mask>> int64(24)))+(((numberIPAddress-1)>> int64(24))&0xFF)))
+	networkRangeQuads = append(networkRangeQuads, fmt.Sprintf("%d", (networkQuads[1]&(s.subnet_mask>> int64(16)))+(((numberIPAddress-1)>> int64(16))&0xFF)))
+	networkRangeQuads = append(networkRangeQuads, fmt.Sprintf("%d", (networkQuads[2]&(s.subnet_mask>> int64(8)))+(((numberIPAddress-1)>> int64(8))&0xFF)))
+	networkRangeQuads = append(networkRangeQuads, fmt.Sprintf("%d", (networkQuads[3]&(s.subnet_mask>> int64(0)))+(((numberIPAddress-1)>> int64(0))&0xFF)))
 
 	return strings.Join(networkRangeQuads, ".")
 }
